% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_match.R
\name{compare_match}
\alias{compare_match}
\title{Check for match and unmatched variable elements, including strings}
\usage{
compare_match(var1, var2)
}
\arguments{
\item{var1}{first variable to compare}

\item{var2}{- second variable}
}
\description{
A function to check for matched and unmatched elements between two vector and returns a list with both matched and unmatched elements.
E.g use match <- (compare_match(var1,var2))$Matched ,to obtain matched elements. which can also be converted to vector using as.vector.
}
