% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatterplot.R
\name{gg_scatter}
\alias{gg_scatter}
\title{Flexible Scatter Plot Wrapper for ggplot2 extensions.}
\usage{
gg_scatter(
  data,
  x,
  y,
  fill = NULL,
  size = NULL,
  alpha = 1,
  shape = 21,
  color = "white",
  cor_test = FALSE,
  size_n = 3,
  fill_n = 3,
  facet_wrap = FALSE,
  wrap_with = NULL,
  ncol = 3,
  scale = "free",
  facet_legend = FALSE,
  ...
)
}
\arguments{
\item{data}{: Dataset (data frame or tibble)}

\item{x}{: a variable name to appear on the x axis (must be numeric).}

\item{y}{: a variable name to appear on the y axis (must be numeric).}

\item{fill}{: Variable to map to point fill color (optional).}

\item{alpha}{: Controls for point transparency.}

\item{color}{: Controls for point border color, and transparency.}

\item{cor_test}{: If TRUE, adds a Pearson correlation coefficient to the plot}

\item{size_n}{: Default size when size is not specified.}

\item{fill_n}{: Default fill when fill color is not specified.}

\item{facet_wrap}{: If TRUE, facets the plot using a categorical variable}

\item{wrap_with}{: Variable used to facet the plot.}

\item{ncol, }{scale: Faceting controls.}

\item{...}{: Additional arguments passed}

\item{size:}{Variable to map to point size (optional).}

\item{shape:}{Controls for point shape, border color, and transparency.}

\item{facet_legend:}{Whether to show the legend in faceted plots}
}
\description{
The gg_scatter() function is a custom wrapper around ggplot2 that generates a scatter plot with optional features such as color fill, point size, correlation test, and faceting. It is designed to be flexible and user-friendly for exploratory data analysis.,
}
\examples{
\dontrun{
if (requireNamespace("gapminder", quietly = TRUE)) {
  library(gapminder)
  library(ggplot2)

  # Basic scatter plot
  gg_scatter(gapminder, x = gdpPercap, y = lifeExp)

  # With fill color
  gg_scatter(gapminder, x = gdpPercap, y = lifeExp, fill = continent)

  # With size and fill
  gg_scatter(gapminder, x = gdpPercap, y = lifeExp, fill = continent, size = pop)

  # With correlation test
  gg_scatter(gapminder, x = gdpPercap, y = lifeExp, cor_test = TRUE)

  # With faceting
  gg_scatter(gapminder, x = gdpPercap, y = lifeExp, fill = continent,
             facet_wrap = TRUE, wrap_with = continent, facet_legend = TRUE)
}
}
}
