% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_fct_label.R
\name{haven_fct_label}
\alias{haven_fct_label}
\title{Extract Haven labels and convert the variable to factor labelled variable.}
\usage{
haven_fct_label(x, ...)
}
\arguments{
\item{x}{haven labelled variable <dbl+lbl>.}

\item{...}{other functions}
}
\description{
The function help to extract haven labels from Stata and SPSS labelled data set/variable, and converts the variable
into factor labeled variable ready for analysis. ie It performs both variable label and value/ levels labels for
the new factor variable,
}
