% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_labels.R
\name{haven_labels}
\alias{haven_labels}
\title{Extract Haven labels.}
\usage{
haven_labels(x, ...)
}
\arguments{
\item{x}{haven labelled variable <dbl+lbl>.}

\item{...}{other functions}
}
\description{
Extract haven labels from Stata and SPSS labelled data set/variable and returns a tibble with values and there labels
}
