% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace _match.R
\name{replace_match}
\alias{replace_match}
\title{Check for match and replace the match with another string word}
\usage{
replace_match(variable, match_text, replace_text)
}
\arguments{
\item{variable}{- the variable to check the match on}

\item{match_text}{- text to find a match for}

\item{replace_text}{- text to replace the match}
}
\description{
Function used to check for a match in a string vector then replace the matched text with a string of interest,
while maintaining the unmatched elements in the variable.
}
