% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratified Systematic sampling-Remmap.R
\name{strat_syst_sampling}
\alias{strat_syst_sampling}
\title{Stratified Systematic Sample for sampling AIM 3 subjects from women who has gone through the screening process described in Aim 1  in REMAPP}
\usage{
strat_syst_sampling(
  data,
  k = 2,
  enrollment_month = c("May", "June"),
  enrollment_year = 2023,
  start_date = "2023-05-01"
)
}
\arguments{
\item{data}{- The REMAPP data set to be operated on}

\item{k}{- sample interval}

\item{enrollment_month}{- Enrollment month(s) to reach trimester III e.g c("May","June")}

\item{enrollment_year}{- Enrollment Year e.g 2023}

\item{start_date}{- subset the data frame to suite when to start the sampling (only the sample frame of interest)}
}
\description{
Function allows you to sample AIM3 cohort for each trimester cross-sectional
}
